+++ SoulBarriers/Barriers/BarrierTypes/Barrier_Hits_Self.cs
@@ -8,7 +8,7 @@ using SoulBarriers.Packets;
 
 namespace SoulBarriers.Barriers.BarrierTypes {
 	public abstract partial class Barrier {
-		public void ApplyHitAgainstSelf( Vector2? hitAt, int damage, bool syncFromServerOnly ) {
+		public void ApplyHitAgainstSelf( Vector2? hitAt, double damage, bool syncFromServerOnly ) {
 			if( syncFromServerOnly && Main.netMode == NetmodeID.MultiplayerClient ) {
 				return;
 			}
@@ -17,7 +17,7 @@ namespace SoulBarriers.Barriers.BarrierTypes {
 				return;
 			}
 
-			int oldStr = this.Strength;
+			double oldStr = this.Strength;
 
 			this.SetStrength( this.Strength - damage );
